MZTolerance: convert to interface
* This will allow different ways of calculating a m/z tolerance (e.g.
  constant m/z, constant ppm, variable m/z or ppm, m/z-specific, etc).
  The primary motivation is to support the error bars returned by
  PeakInvestigator.
* The original MZTolerance class has been renamed CombinedMZTolerance.
