+++ msdk-featdet/msdk-featdet-msmsdetection/src/test/java/io/github/msdk/featdet/msmsdetection/MsMsDetectionMethodTest.java
@@ -29,6 +29,7 @@ import io.github.msdk.datamodel.rawdata.ChromatographyInfo;
 import io.github.msdk.datamodel.rawdata.MsScan;
 import io.github.msdk.datamodel.rawdata.RawDataFile;
 import io.github.msdk.io.mzml.MzMLFileImportMethod;
+import io.github.msdk.util.CombinedMZTolerance;
 import io.github.msdk.util.MZTolerance;
 import io.github.msdk.util.RTTolerance;
 
@@ -63,7 +64,7 @@ public class MsMsDetectionMethodTest {
         }
 
         // Parameters
-        final MZTolerance mzTolerance = new MZTolerance(0.003, 5.0);
+        final MZTolerance mzTolerance = new CombinedMZTolerance(0.003, 5.0);
         final RTTolerance rtTolerance = new RTTolerance(0.2, false);
         final Double intensityTolerance = 0.10d;
