+++ msdk-featdet/msdk-featdet-targeteddetection/src/test/java/io/github/msdk/featdet/targeteddetection/TargetedDetectionMethodTest.java
@@ -32,6 +32,7 @@ import io.github.msdk.datamodel.rawdata.RawDataFile;
 import io.github.msdk.datamodel.rawdata.SeparationType;
 import io.github.msdk.io.mzml.MzMLFileImportMethod;
 import io.github.msdk.util.ChromatogramUtil;
+import io.github.msdk.util.CombinedMZTolerance;
 import io.github.msdk.util.MZTolerance;
 import io.github.msdk.util.RTTolerance;
 
@@ -85,7 +86,7 @@ public class TargetedDetectionMethodTest {
         ionAnnotations.add(ion3);
 
         // Variables
-        final MZTolerance mzTolerance = new MZTolerance(0.003, 5.0);
+        final MZTolerance mzTolerance = new CombinedMZTolerance(0.003, 5.0);
         final RTTolerance rtTolerance = new RTTolerance(0.2, false);
         final Double intensityTolerance = 0.10d;
         final Double noiseLevel = 5000d;
