+++ msdk-features/msdk-features-filtering/src/test/java/io/github/msdk/features/filtering/FeatureFilterMethodTest.java
@@ -40,6 +40,7 @@ import io.github.msdk.featdet.chromatogramtofeaturetable.ChromatogramToFeatureTa
 import io.github.msdk.featdet.targeteddetection.TargetedDetectionMethod;
 import io.github.msdk.io.mzml.MzMLFileImportMethod;
 import io.github.msdk.io.mztab.MzTabFileImportMethod;
+import io.github.msdk.util.CombinedMZTolerance;
 import io.github.msdk.util.MZTolerance;
 import io.github.msdk.util.RTTolerance;
 
@@ -89,7 +90,7 @@ public class FeatureFilterMethodTest {
         ionAnnotations.add(ion3);
 
         // Variables
-        final MZTolerance mzTolerance = new MZTolerance(0.003, 5.0);
+        final MZTolerance mzTolerance = new CombinedMZTolerance(0.003, 5.0);
         final RTTolerance rtTolerance = new RTTolerance(3, false);
         final Double intensityTolerance = 0.10d;
         final Double noiseLevel = 5000d;
