+++ msdk-features/msdk-features-joinaligner/src/test/java/io/github/msdk/features/joinaligner/JoinAlignerMethodTest.java
@@ -29,6 +29,7 @@ import io.github.msdk.datamodel.featuretables.FeatureTable;
 import io.github.msdk.datamodel.featuretables.FeatureTableColumn;
 import io.github.msdk.datamodel.ionannotations.IonAnnotation;
 import io.github.msdk.io.mztab.MzTabFileImportMethod;
+import io.github.msdk.util.CombinedMZTolerance;
 import io.github.msdk.util.MZTolerance;
 import io.github.msdk.util.RTTolerance;
 
@@ -64,7 +65,7 @@ public class JoinAlignerMethodTest {
         featureTables.add(featureTable2);
 
         // Variables
-        MZTolerance mzTolerance = new MZTolerance(0.003, 5.0);
+        MZTolerance mzTolerance = new CombinedMZTolerance(0.003, 5.0);
         RTTolerance rtTolerance = new RTTolerance(0.1, false);
         int mzWeight = 10;
         int rtWeight = 10;
