+++ msdk-features/msdk-features-row-filtering/src/test/java/io/github/msdk/features/rowfilter/RowFilterMethodTest.java
@@ -37,6 +37,7 @@ import io.github.msdk.featdet.chromatogramtofeaturetable.ChromatogramToFeatureTa
 import io.github.msdk.featdet.targeteddetection.TargetedDetectionMethod;
 import io.github.msdk.features.rowfilter.RowFilterMethod;
 import io.github.msdk.io.mzml.MzMLFileImportMethod;
+import io.github.msdk.util.CombinedMZTolerance;
 import io.github.msdk.util.MZTolerance;
 import io.github.msdk.util.RTTolerance;
 
@@ -86,7 +87,7 @@ public class RowFilterMethodTest {
         ionAnnotations.add(ion3);
 
         // Variables
-        final MZTolerance mzTolerance = new MZTolerance(0.003, 5.0);
+        final MZTolerance mzTolerance = new CombinedMZTolerance(0.003, 5.0);
         final RTTolerance rtTolerance = new RTTolerance(3, false);
         final Double intensityTolerance = 0.10d;
         final Double noiseLevel = 5000d;
@@ -121,7 +122,7 @@ public class RowFilterMethodTest {
         Range<Double> mzRange = Range.closed(300.0, 600.0);
         Range<Double> rtRange = Range.closed(600.0, 900.0); // Seconds
         Range<Double> durationRange = Range.closed(0.0, 47.0); // Seconds
-        final MZTolerance duplicateMzTolerance = new MZTolerance(0.003, 5.0);
+        final MZTolerance duplicateMzTolerance = new CombinedMZTolerance(0.003, 5.0);
         final RTTolerance duplicateRtTolerance = new RTTolerance(0.2, false);
         int minCount = 1;
         int minIsotopes = 1;
